requirements.txt
django==3.0.7
djangorestframework==3.11.0
django-crontab==0.7.1
drf-yasg==1.17.1
pandas==1.0.5
whitenoise==5.1.0
gunicorn==20.0.4
django-environ==0.4.5
django-cors-headers==3.4.0
djangorestframework-csv==2.1.0
xlrd==1.2.0
psycopg2-binary==2.8.6